[BUG] Fixing forecasting works with scheduler
As the forecast inherited from Task previously, it was set to only execute once. This is not correct as there may be more than one forecast in the ForecastTask. As it happened, it did not even execute once as the 'reset' method on the scheduler was called before the task had been run. This has been briefly tested.